--- amulet/api/history/entry_manager/__init__.py
@@ -1,2 +1,0 @@
-from .ram import RAMEntry
-from .disk import DiskEntry
