+++ amulet/api/history/history_manager/__init__.py
@@ -1,2 +1,3 @@
+from .object import ObjectHistoryManager
 from .database import DatabaseHistoryManager
 from .meta import MetaHistoryManager
