+++ amulet/api/history/revision_manager/__init__.py
@@ -1,0 +1,2 @@
+from .ram import RAMRevisionManager
+from .disk import DiskRevisionManager
